Squashed commit of the following:
commit 3d7602126de3afcfdeb2f897320e0a648288abf0
Author: Shane Engelman <shane@mitrealityhack.com>
Date:   Sun Dec 17 12:36:37 2023 -0700

    Check if Resume exists before attempting to delete on Application().delete()

+++ infrastructure/models.py
@@ -226,7 +226,8 @@ class Application(models.Model):
 
     @classmethod
     def post_delete(cls, sender, instance, **kwargs):
-        instance.resume.delete()
+        if instance.resume:
+            instance.resume.delete()
 
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     first_name = models.CharField(max_length=100, blank=False, null=False)
